validate types

+++ PGLib/src/patchers/PatcherMeshShaderTruePBR.cpp
@@ -655,21 +655,21 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     const bool enableEnvMapping = truePBRData.contains("env_mapping") && truePBRData["env_mapping"] && !enableTruePBR;
 
     // "delete" attribute
-    if (truePBRData.contains("delete") && truePBRData["delete"]) {
+    if (truePBRData.contains("delete") && truePBRData["delete"].is_boolean() && truePBRData["delete"]) {
         getNIF()->DeleteShape(nifShape);
         changed = true;
         return changed;
     }
 
     // "smooth_angle" attribute
-    if (truePBRData.contains("smooth_angle")) {
+    if (truePBRData.contains("smooth_angle") && truePBRData["smooth_angle"].is_number()) {
         getNIF()->CalcNormalsForShape(nifShape, true, true, truePBRData["smooth_angle"]);
         getNIF()->CalcTangentsForShape(nifShape);
         changed = true;
     }
 
     // "auto_uv" attribute
-    if (truePBRData.contains("auto_uv")) {
+    if (truePBRData.contains("auto_uv") && truePBRData["auto_uv"].is_boolean()) {
         vector<Triangle> tris;
         nifShape->GetTriangles(tris);
         auto newUVScale = autoUVScale(getNIF()->GetUvsForShape(nifShape), getNIF()->GetVertsForShape(nifShape), tris)
@@ -678,7 +678,7 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "vertex_colors" attribute
-    if (truePBRData.contains("vertex_colors")) {
+    if (truePBRData.contains("vertex_colors") && truePBRData["vertex_colors"].is_boolean()) {
         auto newVertexColors = truePBRData["vertex_colors"].get<bool>();
         if (nifShape->HasVertexColors() != newVertexColors) {
             nifShape->SetVertexColors(newVertexColors);
@@ -693,7 +693,8 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
 
     // "vertex_color_lum_mult" and "vertex_color_sat_mult" attribute
     if (nifShape->HasVertexColors()
-        && (truePBRData.contains("vertex_color_lum_mult") || truePBRData.contains("vertex_color_sat_mult"))) {
+        && ((truePBRData.contains("vertex_color_lum_mult") && truePBRData["vertex_color_lum_mult"].is_number())
+            || truePBRData.contains("vertex_color_sat_mult") && truePBRData["vertex_color_sat_mult"].is_number())) {
         vector<BSVertexData>* vertData = nullptr;
         if (dynamic_cast<nifly::BSTriShape*>(nifShape) != nullptr) {
             vertData = &dynamic_cast<nifly::BSTriShape*>(nifShape)->vertData;
@@ -743,13 +744,13 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "zbuffer_write" attribute
-    if (truePBRData.contains("zbuffer_write")) {
+    if (truePBRData.contains("zbuffer_write") && truePBRData["zbuffer_write"].is_boolean()) {
         auto newZBufferWrite = truePBRData["zbuffer_write"].get<bool>();
         changed |= PGNIFUtil::configureShaderFlag(nifShaderBSLSP, SLSF2_ZBUFFER_WRITE, newZBufferWrite);
     }
 
     // "specular_level" attribute
-    if (truePBRData.contains("specular_level")) {
+    if (truePBRData.contains("specular_level") && truePBRData["specular_level"].is_number()) {
         auto newSpecularLevel = truePBRData["specular_level"].get<float>();
         if (nifShader->GetGlossiness() != newSpecularLevel) {
             nifShader->SetGlossiness(newSpecularLevel);
@@ -758,7 +759,8 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "subsurface_color" attribute
-    if (truePBRData.contains("subsurface_color") && truePBRData["subsurface_color"].size() >= 3) {
+    if (truePBRData.contains("subsurface_color") && truePBRData["subsurface_color"].is_array()
+        && truePBRData["subsurface_color"].size() >= 3) {
         auto newSpecularColor = Vector3(truePBRData["subsurface_color"][0].get<float>(),
                                         truePBRData["subsurface_color"][1].get<float>(),
                                         truePBRData["subsurface_color"][2].get<float>());
@@ -769,7 +771,7 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "roughness_scale" attribute
-    if (truePBRData.contains("roughness_scale")) {
+    if (truePBRData.contains("roughness_scale") && truePBRData["roughness_scale"].is_number()) {
         auto newRoughnessScale = truePBRData["roughness_scale"].get<float>();
         if (nifShader->GetSpecularStrength() != newRoughnessScale) {
             nifShader->SetSpecularStrength(newRoughnessScale);
@@ -778,13 +780,13 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "subsurface_opacity" attribute
-    if (truePBRData.contains("subsurface_opacity")) {
+    if (truePBRData.contains("subsurface_opacity") && truePBRData["subsurface_opacity"].is_number()) {
         auto newSubsurfaceOpacity = truePBRData["subsurface_opacity"].get<float>();
         changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->softlighting, newSubsurfaceOpacity);
     }
 
     // "displacement_scale" attribute
-    if (truePBRData.contains("displacement_scale")) {
+    if (truePBRData.contains("displacement_scale") && truePBRData["displacement_scale"].is_number()) {
         auto newDisplacementScale = truePBRData["displacement_scale"].get<float>();
         changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->rimlightPower, newDisplacementScale);
     }
@@ -809,7 +811,7 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "emmissive_scale" attribute
-    if (truePBRData.contains("emissive_scale")) {
+    if (truePBRData.contains("emissive_scale") && truePBRData["emissive_scale"].is_number()) {
         auto newEmissiveScale = truePBRData["emissive_scale"].get<float>();
         if (nifShader->GetEmissiveMultiple() != newEmissiveScale) {
             nifShader->SetEmissiveMultiple(newEmissiveScale);
@@ -818,7 +820,8 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "emmissive_color" attribute
-    if (truePBRData.contains("emissive_color") && truePBRData["emissive_color"].size() >= 4) {
+    if (truePBRData.contains("emissive_color") && truePBRData["emissive_color"].is_array()
+        && truePBRData["emissive_color"].size() >= 4) {
         auto newEmissiveColor = Color4(truePBRData["emissive_color"][0].get<float>(),
                                        truePBRData["emissive_color"][1].get<float>(),
                                        truePBRData["emissive_color"][2].get<float>(),
@@ -830,7 +833,7 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape,
     }
 
     // "uv_scale" attribute
-    if (truePBRData.contains("uv_scale")) {
+    if (truePBRData.contains("uv_scale") && truePBRData["uv_scale"].is_number()) {
         auto newUVScale = Vector2(truePBRData["uv_scale"].get<float>(), truePBRData["uv_scale"].get<float>());
         changed |= PGNIFUtil::setShaderVec2(nifShaderBSLSP->uvScale, newUVScale);
     }
@@ -865,7 +868,8 @@ void PatcherMeshShaderTruePBR::applyOnePatchSlots(PGTypes::TextureSet& slots,
     }
 
     // "emissive" attribute
-    if (truePBRData.contains("emissive") && !flag(truePBRData, "lock_emissive")) {
+    if (truePBRData.contains("emissive") && truePBRData["emissive"].is_boolean()
+        && !flag(truePBRData, "lock_emissive")) {
         wstring newGlow;
         if (truePBRData["emissive"].get<bool>()) {
             newGlow = matchedPath + L"_g.dds";
@@ -875,7 +879,8 @@ void PatcherMeshShaderTruePBR::applyOnePatchSlots(PGTypes::TextureSet& slots,
     }
 
     // "parallax" attribute
-    if (truePBRData.contains("parallax") && !flag(truePBRData, "lock_parallax")) {
+    if (truePBRData.contains("parallax") && truePBRData["parallax"].is_boolean()
+        && !flag(truePBRData, "lock_parallax")) {
         wstring newParallax;
         if (truePBRData["parallax"].get<bool>()) {
             newParallax = matchedPath + L"_p.dds";
@@ -885,7 +890,7 @@ void PatcherMeshShaderTruePBR::applyOnePatchSlots(PGTypes::TextureSet& slots,
     }
 
     // "cubemap" attribute
-    if (truePBRData.contains("cubemap") && !flag(truePBRData, "lock_cubemap")) {
+    if (truePBRData.contains("cubemap") && truePBRData["cubemap"].is_string() && !flag(truePBRData, "lock_cubemap")) {
         auto newCubemap = StringUtil::utf8toUTF16(truePBRData["cubemap"].get<string>());
         slots[static_cast<size_t>(PGEnums::TextureSlots::CUBEMAP)] = newCubemap;
     } else {
@@ -932,7 +937,7 @@ void PatcherMeshShaderTruePBR::applyOnePatchSlots(PGTypes::TextureSet& slots,
         std::string slotName("slot");
         slotName += std::to_string(i + 1);
 
-        if (truePBRData.contains(slotName)) {
+        if (truePBRData.contains(slotName) && truePBRData[slotName].is_string()) {
             std::string newSlot = truePBRData[slotName].get<std::string>();
             StringUtil::toLowerASCIIFastInPlace(newSlot);
 
@@ -957,7 +962,7 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
     applyOnePatchSlots(newSlots, truePBRData, matchedPath);
 
     // "emissive" attribute
-    if (truePBRData.contains("emissive")) {
+    if (truePBRData.contains("emissive") && truePBRData["emissive"].is_boolean()) {
         changed |= PGNIFUtil::configureShaderFlag(
             nifShaderBSLSP, SLSF1_EXTERNAL_EMITTANCE, truePBRData["emissive"].get<bool>());
     }
@@ -976,7 +981,7 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
     changed |= PGNIFUtil::clearShaderFlag(nifShaderBSLSP, SLSF1_EYE_ENVIRONMENT_MAPPING);
 
     // "subsurface" attribute
-    if (truePBRData.contains("subsurface")) {
+    if (truePBRData.contains("subsurface") && truePBRData["subsurface"].is_boolean()) {
         changed |= PGNIFUtil::configureShaderFlag(
             nifShaderBSLSP, SLSF2_RIM_LIGHTING, truePBRData["subsurface"].get<bool>());
     }
@@ -988,7 +993,8 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
 
     // "multilayer" attribute
     bool enableMultiLayer = false;
-    if (truePBRData.contains("multilayer") && truePBRData["multilayer"]) {
+    if (truePBRData.contains("multilayer") && truePBRData["multilayer"].is_boolean()
+        && truePBRData["multilayer"].get<bool>()) {
         enableMultiLayer = true;
 
         changed |= PGNIFUtil::setShaderType(nifShader, BSLSP_MULTILAYERPARALLAX);
@@ -1006,48 +1012,48 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
         }
 
         // "coat_specular_level" attribute
-        if (truePBRData.contains("coat_specular_level")) {
+        if (truePBRData.contains("coat_specular_level") && truePBRData["coat_specular_level"].is_number()) {
             auto newCoatSpecularLevel = truePBRData["coat_specular_level"].get<float>();
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxRefractionScale, newCoatSpecularLevel);
         }
 
         // "coat_roughness" attribute
-        if (truePBRData.contains("coat_roughness")) {
+        if (truePBRData.contains("coat_roughness") && truePBRData["coat_roughness"].is_number()) {
             auto newCoatRoughness = truePBRData["coat_roughness"].get<float>();
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxInnerLayerThickness, newCoatRoughness);
         }
 
         // "coat_strength" attribute
-        if (truePBRData.contains("coat_strength")) {
+        if (truePBRData.contains("coat_strength") && truePBRData["coat_strength"].is_number()) {
             auto newCoatStrength = truePBRData["coat_strength"].get<float>();
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->softlighting, newCoatStrength);
         }
 
         // "coat_diffuse" attribute
-        if (truePBRData.contains("coat_diffuse")) {
+        if (truePBRData.contains("coat_diffuse") && truePBRData["coat_diffuse"].is_boolean()) {
             changed |= PGNIFUtil::configureShaderFlag(
                 nifShaderBSLSP, SLSF2_EFFECT_LIGHTING, truePBRData["coat_diffuse"].get<bool>());
         }
 
         // "coat_parallax" attribute
-        if (truePBRData.contains("coat_parallax")) {
+        if (truePBRData.contains("coat_parallax") && truePBRData["coat_parallax"].is_boolean()) {
             changed |= PGNIFUtil::configureShaderFlag(
                 nifShaderBSLSP, SLSF2_SOFT_LIGHTING, truePBRData["coat_parallax"].get<bool>());
         }
 
         // "coat_normal" attribute
-        if (truePBRData.contains("coat_normal")) {
+        if (truePBRData.contains("coat_normal") && truePBRData["coat_normal"].is_boolean()) {
             changed |= PGNIFUtil::configureShaderFlag(
                 nifShaderBSLSP, SLSF2_BACK_LIGHTING, truePBRData["coat_normal"].get<bool>());
         }
 
         // "inner_uv_scale" attribute
-        if (truePBRData.contains("inner_uv_scale")) {
+        if (truePBRData.contains("inner_uv_scale") && truePBRData["inner_uv_scale"].is_number()) {
             auto newInnerUVScale
                 = Vector2(truePBRData["inner_uv_scale"].get<float>(), truePBRData["inner_uv_scale"].get<float>());
             changed |= PGNIFUtil::setShaderVec2(nifShaderBSLSP->parallaxInnerLayerTextureScale, newInnerUVScale);
         }
-    } else if (truePBRData.contains("glint")) {
+    } else if (truePBRData.contains("glint") && truePBRData["glint"].is_object()) {
         // glint is enabled
         const auto& glintParams = truePBRData["glint"];
 
@@ -1057,26 +1063,26 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
         changed |= PGNIFUtil::setShaderFlag(nifShaderBSLSP, SLSF2_FIT_SLOPE);
 
         // Glint parameters
-        if (glintParams.contains("screen_space_scale")) {
+        if (glintParams.contains("screen_space_scale") && glintParams["screen_space_scale"].is_number()) {
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxInnerLayerThickness,
                                                  glintParams["screen_space_scale"]);
         }
 
-        if (glintParams.contains("log_microfacet_density")) {
+        if (glintParams.contains("log_microfacet_density") && glintParams["log_microfacet_density"].is_number()) {
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxRefractionScale,
                                                  glintParams["log_microfacet_density"]);
         }
 
-        if (glintParams.contains("microfacet_roughness")) {
+        if (glintParams.contains("microfacet_roughness") && glintParams["microfacet_roughness"].is_number()) {
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxInnerLayerTextureScale.u,
                                                  glintParams["microfacet_roughness"]);
         }
 
-        if (glintParams.contains("density_randomization")) {
+        if (glintParams.contains("density_randomization") && glintParams["density_randomization"].is_number()) {
             changed |= PGNIFUtil::setShaderFloat(nifShaderBSLSP->parallaxInnerLayerTextureScale.v,
                                                  glintParams["density_randomization"]);
         }
-    } else if (truePBRData.contains("fuzz")) {
+    } else if (truePBRData.contains("fuzz") && truePBRData["fuzz"].is_object()) {
         // fuzz is enabled
         const auto& fuzzParams = truePBRData["fuzz"];
 
@@ -1087,7 +1093,7 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
 
         // get color
         auto fuzzColor = vector<float> {0.0F, 0.0F, 0.0F};
-        if (fuzzParams.contains("color")) {
+        if (fuzzParams.contains("color") && fuzzParams["color"].is_array()) {
             fuzzColor = fuzzParams["color"].get<vector<float>>();
         }
 
@@ -1097,7 +1103,7 @@ auto PatcherMeshShaderTruePBR::enableTruePBROnShape(NiShader* nifShader,
 
         // get weight
         auto fuzzWeight = 1.0F;
-        if (fuzzParams.contains("weight")) {
+        if (fuzzParams.contains("weight") && fuzzParams["weight"].is_number()) {
             fuzzWeight = fuzzParams["weight"].get<float>();
         }
 
@@ -1159,5 +1165,5 @@ auto PatcherMeshShaderTruePBR::autoUVScale(const vector<Vector2>* uvs,
 auto PatcherMeshShaderTruePBR::flag(const nlohmann::json& json,
                                     const char* key) -> bool
 {
-    return json.contains(key) && json[key];
+    return json.contains(key) && json[key].is_boolean() && json[key];
 }
